networking/requirements.txt
websocket-client==1.4.1
websockets==10.3